refactor(#88): WebCorsConfig WebConfig로 통합
